Made main menu have more consistent UI.
Made all buttons have the same sprite, font color/style. A good addition to this would be to make it on hover, swap the button's sprite with buttons_13 to make it slightly darker, like it's been pressed in.
Also made them all use the same scale with regard to the main menu object.