Made options panel like advanced options panel
Made the sliders and proximity have parent objects to make centering easier. Also, made the dropdown text a different shade of blue. Even though it was the same as the rest of the test, since it was smaller, it looked black, so I lightened it and it looks about the same as the rest of the text.