Adjusted Game Options Script to handle a custom scene
Made script activate and deactivate custom objects that the player will be able to interact with if they choose to have a custom version of the game.
Next step is to make a drop down for users to select prey, predator, and background.
Once that is done, make those selections be sent into the custom scene...